Bug fix: process.version parsing error
This fixes a `process.version` parsing bug introduced in
65d1d5138b0733b5a10de0cb459dab0d1fcdc109.

`process.version returns a version string in the format v0.12.0 rather
`than 0.12.0. Since the first token from split returns a string v0
`rather than 0, map function always maps it to NaN. As a result
`version[0] === 0 test always fails and process.nextTick is always
`picked.

The suggested fix is to use `process.versions.node` as it returns a
string in the format `0.12.0`. (also confirmed this value exists and
works as expected in `io.js`.